changes to allow for different dtypes
it can be helpful to use float16 to save memory. if that's the case,
most things work as you'd hope with just a .to(torch.float16) call,
but need to make sure that to_numpy makes it a float32 and that we get
our data on the same dtype (in addition to device) as PoolingWindows
in plot_representation_image

+++ plenoptic/tools/data.py
@@ -12,7 +12,7 @@ DATA_PATH = op.join(op.dirname(op.realpath(__file__)), '..', '..', 'data')
 def to_numpy(x):
     r"""cast tensor to numpy in the most conservative way possible
     """
-    return x.detach().cpu().numpy()
+    return x.detach().cpu().numpy().astype(np.float32)
 
 
 def make_basic_stimuli(size=256, requires_grad=True):
